fix: seed columns must be BIGINT (Postgres INTEGER is 32-bit)
Deterministic seeds are 63-bit; SQLite INTEGER holds them, Postgres INTEGER
overflows -> full-song generation failed on the deployed DB with
'integer out of range'. song/section/job/render-take/assignment seed columns
-> BigInteger, + migration 5c3232296ddf. 140 tests pass.

+++ sr/models/generation_job.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import JSON, DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import JSON, BigInteger, DateTime, Float, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from sr.models.base import Base, Timestamps, UUIDPrimaryKey
@@ -30,7 +30,7 @@ class GenerationJob(UUIDPrimaryKey, Timestamps, Base):
     status: Mapped[str] = mapped_column(String(20), default="queued", index=True)
     progress: Mapped[float] = mapped_column(Float, default=0.0)  # 0..1
 
-    seed: Mapped[int | None] = mapped_column(Integer, default=None)
+    seed: Mapped[int | None] = mapped_column(BigInteger, default=None)
     parameters_json: Mapped[dict | None] = mapped_column(JSON, default=None)
     result_json: Mapped[dict | None] = mapped_column(JSON, default=None)
     input_asset_ids: Mapped[list | None] = mapped_column(JSON, default=None)
